NOT_AVAILABLE_IN_STATE(10506): text update
This error is used currently only in web UI when one-click should be shown but printer is not ready to print. In this case wizard is started so user cannot start a print from web UI. This is better description recommending to the user to check the printer physically and finish the wizard.